fix(core): escape the Unicode line separators quoting alone leaves literal
The run-state rejection quoted the id it refused with JSON.stringify, which
escapes a newline but leaves U+0085, U+2028 and U+2029 in place. A JavaScript
`^` under `/m` breaks on the latter two, so a refused run id could still put a
forged control block at a line start, and a reconcile rethrows that message
without passing it through the init path's sanitizer.

The collapsing helper the emit sites already used covers all three, so the
diagnostic now collapses before it quotes. It moves to its own module to get
there: run-state.ts cannot reach util.ts, whose own import of state-lock
closes a cycle back onto run-state. Nothing else about the helper changes, and
every renderer of agent-facing text now shares the one definition.
